% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/export_stage.r
\name{build_export_stagerunner}
\alias{build_export_stagerunner}
\title{Build a stagerunner for exporting data with backup sources.}
\usage{
build_export_stagerunner(export_options)
}
\arguments{
\item{export_options}{list. Nested list, one adapter per list entry.
These adapter parametrizations will get converted to legitimate
IO adapters. (See the "adapter" reference class.)}
}
\description{
Build a stagerunner for exporting data with backup sources.
}

