% Generated by roxygen2 (4.0.0): do not edit by hand
\name{build_import_stagerunner}
\alias{build_import_stagerunner}
\title{Build a stagerunner for importing data with backup sources.}
\usage{
build_import_stagerunner(import_options)
}
\arguments{
\item{import_options}{list. Nested list, one adapter per list entry.
These adapter parametrizations will get converted to legitimate
IO adapters. (See the "adapter" reference class.)}
}
\description{
Build a stagerunner for importing data with backup sources.
}

