% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adapter.r
\name{common_s3mpi_package_loader}
\alias{common_s3mpi_package_loader}
\title{Check if s3mpi package is installed and loaded.}
\usage{
common_s3mpi_package_loader()
}
\value{
\code{TRUE} or \code{FALSE} indicating if loading and
 attaching is successful.
}
\description{
Stopped if s3mpi package is not installed.
}

