% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluation_stage.r
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Plot a confusion matrix for a given prediction set, and return the table.}
\usage{
confusion_matrix(dataframe, cutoff = 0.5, plot.it = TRUE,
  xlab = c("dep_var = 0", "dep_var = 1"), ylab = c("score = 0",
  "score = 1"), title = NULL)
}
\arguments{
\item{dataframe}{data.frame. Must contain \code{score} and \code{dep_var}
columns. The confusion matrix will be calculated for these values.
The mentioned columns must both be numeric.}

\item{cutoff}{numeric. The cutoff at which to assign numbers greater a 1
for prediction purposes, and 0 otherwise. The default is 0.5.}

\item{plot.it}{logical. Whether or not to plot the confusion matrix as a
four fold diagram. The default is \code{TRUE}.}

\item{xlab}{character. The labels for the rows (\code{dep_var}). The default
is \code{c("dep_var = 0", "dep_var = 1")}.}

\item{ylab}{character. The labels for the rows (\code{score}). The default
is \code{c("score = 0", "score = 1")}.}

\item{title}{character. The title for the fourfoldplot, if it is graphed.}
}
\value{
a table. The confusion matrix table.
}
\description{
Plot a confusion matrix for a given prediction set, and return the table.
}

