% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/evaluation_stage.r
\name{evaluation_stage_auc}
\alias{evaluation_stage_auc}
\title{Compute the AUC in conjunction with \code{evaluation_stage}.}
\usage{
evaluation_stage_auc(modelenv)
}
\arguments{
\item{modelenv}{environment. The current modeling environment.}
}
\description{
The RoC, accuracy, sensitivity, and specificy information will be
returned as computed by the AUC package.
}
\details{
The modeling environment will contain a key \code{auc} inside
of the \code{evaluation_stage} list. This will contain:

\itemize{
   \item auc. The auc of the ROC of the validation data.
   \item ci_auc. The 95% confidence interval for the auc of the ROC of the validation data.
   \item accuracy. The accuracy of the validation data.
   \item sensitivity. The sensitivity of the validation data.
   \item specificity. The specificity of the validation data.
}
}

