% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/evaluation_stage.r
\name{evaluation_stage_confusion_matrix}
\alias{evaluation_stage_confusion_matrix}
\title{Compute a confusion matrix and plot it.}
\usage{
evaluation_stage_confusion_matrix(modelenv)
}
\arguments{
\item{modelenv}{environment. The current modeling environment.}
}
\description{
The results will be stored in the modeling environment's
\code{confusion_matrix} key within the \code{evaluation_stage} list.
}

