% Generated by roxygen2 (4.0.0): do not edit by hand
\name{evaluation_stage_generate_options}
\alias{evaluation_stage_generate_options}
\title{Store necessary information for evaluation stage.}
\usage{
evaluation_stage_generate_options(modelenv)
}
\arguments{
\item{modelenv}{environment. The current modeling environment.}
}
\description{
The \code{evaluation_stage} function needs a data.frame containing
some information (namely, \code{score} and \code{dep_var}). This function
extracts that information from the active stageRunner. Note this
means evaluation will not work if not called from syberia's \code{run_model}.
This will later be alleviated by the introduction of data partitions
into syberia.
}
\details{
If the evaluation stage prints a CSV copy of the prediction dataframe
(see the \code{output} parameter in the \code{evaluation_stage} options,
it is also performed in this helper function.
}

