% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluation_stage.r
\name{evaluation_stage_validation_plot}
\alias{evaluation_stage_validation_plot}
\title{Draw a decile validation plot for a classification problem.}
\usage{
evaluation_stage_validation_plot(modelenv)
}
\arguments{
\item{modelenv}{environment. The current modeling environment.}
}
\description{
This evaluation stage will produce a plot showing performance
of the model according to a certain metric described below.
The plot will be stored in a PNG file, which the user can control.
(See the \code{output} option in \code{evaluation_stage}).
}
\details{
The plot will contain the following information:

\itemize{
 \item Percent of validation samples with \code{dep_var} = 1 on the y-axis.
 \item Decile by model score on the x-axis. For example, if 0.1 is displayed
   on the x-axis, this should be interpreted as "the best 10% of samples
   by decile score" (meaning, most likely to have \code{dep_var = 0} as
   predicted by the model). If 0.2 is displayed, it means "the next best 10%
   of samples by decile score", that is, those falling within the 10 to 20th
   percentile.
}

In other words, the plot describes how "good" the model is at capturing
the actual percent of samples with \code{dep_var = 1}. The samples whose
predicted model score is within the top 10% should have low percent of
empirical \code{dep_var = 1}, whereas the worst 10% should have high
such percentage.
}

