% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fetch_adapter}
\alias{fetch_adapter}
\title{Fetch an import adapter.}
\usage{
fetch_adapter(keyword)
}
\arguments{
\item{keyword}{character. The keyword for the adapter (e.g., 'file', 's3', etc.)}
}
\value{
an \code{adapter} object (defined in this package, syberiaStages)
}
\description{
Fetch an import adapter.
}

