% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fetch_custom_adapter}
\alias{fetch_custom_adapter}
\title{Fetch a customer syberia IO adapter.}
\usage{
fetch_custom_adapter(keyword)
}
\arguments{
\item{keyword}{character. The keyword for the adapter (e.g., 'file', 's3', etc.)}
}
\value{
an \code{adapter} object (defined in this package, syberiaStages)
}
\description{
Custom adapters are defined in \code{lib/adapters} from the root
of the syberia project. Placing a file there with, for example, name 'foo.R',
will cause \code{fetch_custom_adapter('foo')} to return an appropriate
IO adapter. The file 'foo.R' must contain a 'read', 'write', and (optionally)
'format' function, which will be used to construct the adapter. (See
the definition of the adapter reference class.)
}

