% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_stage.r
\name{import_stage}
\alias{import_stage}
\title{Import data stage for Syberia model process.}
\usage{
import_stage(import_options)
}
\arguments{
\item{import_options}{list. The available import options. Will differ
depending on the adapter. (default is file adapter)}
}
\description{
Import data stage for Syberia model process.
}

