% Generated by roxygen2 (4.0.0): do not edit by hand
\name{parse_custom_adapter}
\alias{parse_custom_adapter}
\title{Ensures a custom adapter resource is valid and returns the corresponding
adapter reference class object.}
\usage{
parse_custom_adapter(provided_env, type)
}
\arguments{
\item{provided_env}{environment. The environment the adapter was loaded from.}

\item{type}{character. The keyword for the adapter.}
}
\value{
the \code{adapter} reference class object constructed from the parsed
   adapter resource.
}
\description{
There can only be one function defined that contains the string "read".
Similarly there can only be one such function containing "write".
If this condition is not met, this function will throw an error.
Finally, there is also an optional "format" function that can be defined.
}

