% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/model_stage.r
\name{parse_custom_classifier}
\alias{parse_custom_classifier}
\title{Ensures a custom classifier is valid and returns its train and predict methods.}
\usage{
parse_custom_classifier(provided_env, type)
}
\arguments{
\item{provided_env}{environment. The environment the classifier was loaded from.}

\item{type}{character. The keyword for the classifier.}
}
\value{
a list containing keys "train" and "predict" indicating the train
   and predict functions.
}
\description{
There can only be one function defined that contains the string "train".
Similarly there can only be one such function containing "predict".
If this condition is not met, this function will throw an error.
}

