% Generated by roxygen2 (4.0.0): do not edit by hand
\name{parse_custom_functions}
\alias{parse_custom_functions}
\title{Parse functions out of a custom resource.}
\usage{
parse_custom_functions(functions, provided_env, type,
  resource_type = "classifier", strict = TRUE)
}
\arguments{
\item{functions}{character. The names of functions to parse out.}

\item{provided_env}{environment. The environment the resource was loaded from.}

\item{type}{character. The keyword for the resource.}

\item{resource_type}{character. The type of the resource (e.g., "classifier",
"adapter", etc.). This will be used to generate error messages.}

\item{strict}{logical. Whether or not to error if the functions are not found.}
}
\value{
a list containing keys same as the \code{functions} argument.
   and predict functions.
}
\description{
Parse functions out of a custom resource.
}

