% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/syberiaStages-package.r
\docType{package}
\name{syberiaStages}
\alias{syberiaStages}
\alias{syberiaStages-package}
\title{Syberia provides an opinionated unified framework for
fast iteration on classifier development and deployment. It is
founded on convention over configuration and aims to solve the
problems of classifier-specific data preparation and
classifier-specific modeling parameters.}
\description{
syberiaStages is a library of stages for Syberia.
}

