% Generated by roxygen2 (4.0.0): do not edit by hand
\name{common_file_formatter}
\alias{common_file_formatter}
\title{A helper function for formatting parameters for adapters to
correctly include an argument "file", with aliases
"resource", "filename", "name", and "path".}
\usage{
common_file_formatter(opts)
}
\arguments{
\item{opts}{list. The options that will get passed to the adapter
constructor function.}
}
\value{
the fixed and sanitized formatted options.
}
\description{
A helper function for formatting parameters for adapters to
correctly include an argument "file", with aliases
"resource", "filename", "name", and "path".
}

