% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adapter.r
\name{construct_file_adapter}
\alias{construct_file_adapter}
\title{Construct a file adapter.}
\usage{
construct_file_adapter()
}
\value{
an \code{adapter} object which reads and writes to a file.
}
\description{
Construct a file adapter.
}

