% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/adapter.r
\name{construct_R_adapter}
\alias{construct_R_adapter}
\title{Construct an adapter for reading to and from an R environment,
by default the global environment.}
\usage{
construct_R_adapter()
}
\value{
an \code{adapter} object which reads and writes to Amazon's S3.
}
\description{
Construct an adapter for reading to and from an R environment,
by default the global environment.
}

