% Generated by roxygen2 (4.0.0): do not edit by hand
\name{construct_s3_adapter}
\alias{construct_s3_adapter}
\title{Construct an Amazon Web Services S3 adapter.}
\usage{
construct_s3_adapter()
}
\value{
an \code{adapter} object which reads and writes to Amazon's S3.
}
\description{
This requires that the user has set up the s3mpi package to
work correctly (for example, the s3mpi.path option should be set).
(Note that this adapter is not related to R's S3 classes).
}

