% Generated by roxygen2 (4.0.0): do not edit by hand
\name{default_adapter}
\alias{default_adapter}
\title{Fetch the default adapter keyword from the active syberia
project's configuration file.}
\usage{
default_adapter()
}
\value{
a string representing the default adapter.
}
\description{
Fetch the default adapter keyword from the active syberia
project's configuration file.
}

