% Generated by roxygen2 (4.0.0): do not edit by hand
\name{evaluation_stage}
\alias{evaluation_stage}
\title{Evaluation stage for binary classification syberia models.}
\usage{
evaluation_stage(evaluation_parameters)
}
\arguments{
\item{evaluation_parameters}{list. These come from the syberia model and
   must be in the following format:

   \itemize{
     \item output. The prefix of the CSV and PNG to which to output the results
       of the validation. For example, if you put "output/foo", then "output/foo.csv"
       will be a CSV file containing a data.frame with the columns "dep_var",
       "score", and \code{id_column}, where the score refers to predicted score,
       hand, "output/foo.png" will contain a decile validation plot of the results.
     \item percent. The percent of the data that was used for training. Currently,
       only sequential splits of training and validation are supported until
       syberia introduces a better mechanism for data partitions. The default is 0.8.
     \item dep_var. (Optional) The name of the dependent variable in the evaluated data.
       The default is "dep_var".
     \item id_column. (Optional) The name of an identifying column in your
       pre-data-munged data.frame. This will be included in the validation
       output CSV. If not given, no ID column will be included.
     \item cutoff. (Optional) A cutoff for binary classification predictions: above the
       cutoff means a prediction 1, and otherwise a 0. The default is 0.5.
    }}
}
\value{
a stageRunner that performs AUC, confusing matrix, and validation
  plotting.
}
\description{
A helper stage for evaluating a binary classification model
according to the metrics: AUC, a confusion matrix, and a validation plot.
}
\details{
The evaluation stage parameters that can be controlled through the
syberia model file are described in the evaluation_parameters argument.
For example, just like how you can set \code{import = list(...)} to
control what data gets imported in your syberia model, you can
write \code{evaluation = list(output = 'foo', percent = 0.6, ...)} to
control what happens during this evaluation_stage.
}

