% Generated by roxygen2 (4.0.0): do not edit by hand
\name{evaluation_stage}
\alias{evaluation_stage}
\title{Evaluation stage for syberia models.}
\usage{
evaluation_stage(modelenv, evaluation_parameters)
}
\arguments{
\item{modelenv}{an environment. The persistent modeling environment.}

\item{evaluation_parameters}{list. These come from the syberia model and
   must be in the following format:

   \itemize{
     \item type. The first element must be a string "regression" or "classification".
     \item percent. The percent of the data that was used for training. Currently,
       only sequential splits of training and validation are supported until
       syberia introduces a better mechanism for data partitions. The default is 0.8.
    }}
}
\value{
a stageRunner that performs AUC, confusing matrix, and validation
  plotting.
}
\description{
A helper stage for evaluating a binary classification or a regression
model according to the metrics: AUC, a confusion matrix, and a validation plot.
}
\details{
The evaluation stage parameters in the syberia model file are described in the
evaluation_parameters argument.
}

