% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluation_stage.r
\name{evaluation_stage_generate_options}
\alias{evaluation_stage_generate_options}
\title{Store necessary information for evaluation stage.}
\usage{
evaluation_stage_generate_options(params)
}
\arguments{
\item{params}{list. A list containing \code{output}, \code{cutoff},
\code{train_percent}, \code{dep_var}, and \code{id_column} as in
the evaluation_stage parameters.}
}
\value{
a function suitable for use in a stageRunner.
}
\description{
The \code{evaluation_stage} function needs a data.frame containing
some information (namely, \code{score} and \code{dep_var}). This function
extracts that information from the active stageRunner. Note this
means evaluation will not work if not called from syberia's \code{run_model}.
This will later be alleviated by the introduction of data partitions
into syberia.
}
\details{
If the evaluation stage prints a CSV copy of the prediction dataframe
(see the \code{output} parameter in the \code{evaluation_stage} options,
it is also performed in this helper function.
}

