% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/export_stage.r
\name{export_stage}
\alias{export_stage}
\title{Export data stage for Syberia model process.}
\usage{
export_stage(export_options)
}
\arguments{
\item{export_options}{list. The available export options. Will differ
depending on the adapter. (default is file adapter)}
}
\description{
Export data stage for Syberia model process.
}

