% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fetch_adapter}
\alias{fetch_adapter}
\title{Fetch a syberia IO adapter.}
\usage{
fetch_adapter(keyword)
}
\arguments{
\item{keyword}{character. The keyword for the adapter (e.g., 'file', 's3', etc.)}
}
\value{
an \code{adapter} object (defined in this package, syberiaStages)
}
\description{
IO adapters are (reference class) objects that have a \code{read}
and \code{write} method. By wrapping things in an adapter, you do not have to
worry about whether to use, e.g., \code{read.csv} versus \code{s3read}
or \code{write.csv} versus \code{s3store}. If you are familiar with
the tundra package, think of adapters as like tundra containers for
importing and exporting data.
}
\details{
For example, we can do: \code{fetch_adapter('file')$write(iris, '/tmp/iris.csv')}
and the contents of the built-in \code{iris} data set will be stored
in the file \code{"/tmp/iris.csv"}.
}

