% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adapter.r
\name{fetch_custom_adapter}
\alias{fetch_custom_adapter}
\title{Fetch a custom syberia IO adapter.}
\usage{
fetch_custom_adapter(keyword, modified_check = FALSE)
}
\arguments{
\item{keyword}{character. The keyword for the adapter (e.g., 'file', 's3', etc.)}

\item{modified_check}{logical. If \code{TRUE}, will return a logical indicating
whether or not the customer adapter has been modified. By default, \code{FALSE}.}
}
\value{
an \code{adapter} object (defined in this package, syberiaStages)
}
\description{
Custom adapters are defined in \code{lib/adapters} from the root
of the syberia project. Placing a file there with, for example, name 'foo.R',
will cause \code{fetch_custom_adapter('foo')} to return an appropriate
IO adapter. The file 'foo.R' must contain a 'read', 'write', and (optionally)
'format' function, which will be used to construct the adapter. (See
the definition of the adapter reference class.)
}

