% Generated by roxygen2 (4.0.0): do not edit by hand
\name{fetch_syberia_adapter}
\alias{fetch_syberia_adapter}
\title{Publically exported version of \code{fetch_adapter}.}
\usage{
fetch_syberia_adapter(keyword)
}
\arguments{
\item{keyword}{character. The keyword for the adapter (e.g., 'file', 's3', etc.)}
}
\description{
Publically exported version of \code{fetch_adapter}.
}
\seealso{
\code{\link{fetch_adapter}}
}

